Rename helpers.js
